feat(StationController): check station name length
