+++ docs/README.md
@@ -7,7 +7,7 @@
     - [x] 종료가 선택된 경우 프로그램을 종료한다.
   
 - [x] 역 이름을 입력받는다.
-    - [ ] 역 이름이 2글자 이상인지 확인한다.
+    - [x] 역 이름이 2글자 이상인지 확인한다.
     - [ ] '역'으로 끝나는 이름인지 확인한다.
     - [ ] 역이 이미 존재하는지 확인한다.
     - [ ] 역을 지우는 경우 그 역이 노선에 포함되어 있는지 확인한다.
