--- src/main/java/subway/domain/Station.java
@@ -1,15 +1,0 @@
-package subway.domain;
-
-public class Station {
-    private String name;
-
-    public Station(String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    // 추가 기능 구현
-}
